services/metric-service/metric-api/src/pyproject.toml
[project]
name = "metric-service"
version = "0.1.0"
description = "Authentication microservice"
requires-python = ">=3.12,<4.0"
dependencies = [
    "flask >=3.1.1,<4.0.0",
    "packaging (>=25.0,<26.0)",
    "kafka-python (>=2.2.10,<3.0.0)",
    "backoff (>=2.2.1,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pydantic (>=2.11.5,<3.0.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"
